Styles added to the cats pages.
The design is okay-ish if we are being honest, then again - it was done in a few days.

The point isn't to make a design marvel - but to get used to React Router. Infact, this project might not even receive mobile/tablet support for that reason. Design is not a priority at all here.